Rename wizardAuth
