+++ src/index.js
@@ -5,7 +5,7 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter as Router } from "react-router-dom";
 import { Provider } from "react-redux";
-import Store from "./Store";
+import Store from "./store";
 import { ThemeProvider } from "react-jss";
 
 const theme = {
